add config files to load in models instead of hardcoded values
