+++ model-config/gemini/gemini-config.ini
@@ -1,0 +1,8 @@
+[gemini-1.0-pro]
+context_length = 32000
+
+[gemini-1.5-pro]
+context_length = 2000000
+
+[gemini-1.5-flash]
+context_length = 1000000
