+++ model-config/ollama/ollama-config.ini
@@ -1,32 +1,3 @@
-# Any new models that someone wants to integrate should be added here
-
-# gpt models
-[gpt-3.5-turbo]
-context_length = 16385
-
-[gpt-3.5-turbo-1106]
-context_length = 16385
-
-[gpt-4-turbo-preview]
-context_length = 128000
-
-[gpt-4-turbo]
-context_length = 128000
-
-[gpt-4o-mini]
-context_length = 128000
-
-# gemini models
-[gemini-pro]
-context_length = 32000
-
-[gemini-1.5-pro]
-context_length = 2000000
-
-[gemini-1.5-flash]
-context_length = 1000000
-
-# ollama models
 [llama3.1]
 context_length = 128000
 model_id = meta-llama/Meta-Llama-3.1-8B-Instruct
