+++ model-config/openai/openai-config.ini
@@ -1,0 +1,16 @@
+[gpt-3.5-turbo]
+context_length = 16385
+
+[gpt-3.5-turbo-1106]
+context_length = 16385
+
+[gpt-4-turbo-preview]
+context_length = 128000
+
+[gpt-4-turbo]
+context_length = 128000
+
+[gpt-4o-mini]
+context_length = 128000
+
+# Add any extra OpenAI models here
